fix(tasks): department_head redirige a Mi Equipo al login + SQL BIGINT
- Login redirige department_head a /hrm/department-status en vez de
  /blank-home (el encargado ve su panel directamente al entrar)
- Migration SQL actualizado: designation_id y scoped_designation_id
  ahora usan BIGINT (consistente con producción)

Auditoría de consistencia completa:
✅ Field names: TaskModel, DesignationModel, EmployeeModel correctos
✅ Routes: 4 rutas HRM + permission mappings verificados
✅ API paths: task_repo.dart coincide con backend
✅ Sidebar paths: coinciden con routes
✅ Null safety: department_status cards con null guards
✅ Provider refresh: reservationsProvider se refresca al asignar
✅ Timer/AnimationController: dispose correcto en countdown card

+++ migration/tasks_system_migration.sql
@@ -51,7 +51,7 @@ CREATE INDEX IF NOT EXISTS idx_employees_user_id ON {{SCHEMA}}.employees(user_id
 -- ============================================================
 
 ALTER TABLE public.users
-  ADD COLUMN IF NOT EXISTS scoped_designation_id INT,
+  ADD COLUMN IF NOT EXISTS scoped_designation_id BIGINT,
   ADD COLUMN IF NOT EXISTS linked_employee_id UUID;
 
 COMMENT ON COLUMN public.users.scoped_designation_id IS 'Si está presente, el user solo ve datos de esta designación (encargado o empleado)';
@@ -64,7 +64,7 @@ COMMENT ON COLUMN public.users.linked_employee_id IS 'Si el user representa a un
 CREATE TABLE IF NOT EXISTS {{SCHEMA}}.tasks (
   id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
   reservation_id UUID NOT NULL,
-  designation_id INT NOT NULL,
+  designation_id BIGINT NOT NULL,
   assigned_to_user_id UUID,
   assigned_to_employee_id UUID,
   assigned_by_user_id UUID,
